Made it more modular
